abstract resource
